requirements.txt
numpy>=1.24.0
numba>=0.57.0
rasterio>=1.3.0
matplotlib>=3.7.0
PyQt6>=6.5.0
scipy>=1.10.0
pyinstaller>=5.13.0
pytest>=7.4.0
pyyaml>=6.0.0
pyvista>=0.42.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.0
